+++ database/factories/AuthenticationLogFactory.php
@@ -5,14 +5,13 @@ declare(strict_types=1);
 namespace Modules\User\Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use Modules\User\Models\AuthenticationLog;
 
 class AuthenticationLogFactory extends Factory
 {
     /**
      * The name of the factory's corresponding model.
      */
-    protected $model = AuthenticationLog::class;
+    protected $model = \Modules\User\Models\AuthenticationLog::class;
 
     /**
      * Define the model's default state.