+++ database/migrations/2024_09_26_100442_create_features_table.php
@@ -15,21 +15,16 @@ return new class extends XotBaseMigration {
      */
     public function up(): void
     {
-        // -- CREATE --
-        $this->tableCreate(static function (Blueprint $table): void {
-            $table->id();
-            $table->string('name');
-            $table->string('scope');
-            $table->text('value');
-
-            $table->unique(['name', 'scope']);
-        });
-        // -- UPDATE --
-        $this->tableUpdate(function (Blueprint $table): void {
-            $this->updateTimestamps(
-                table: $table,
-                hasSoftDeletes: true,
-            );
-        });
+        if (! $this->tableExists()) {
+            $this->tableCreate(static function (Blueprint $table): void {
+                $table->id();
+                $table->string('name');
+                $table->string('scope');
+                $table->text('value');
+                $table->unique(['name', 'scope']);
+                $table->timestamps(); // Add timestamps here
+                $table->softDeletes(); // Add soft deletes here, as hasSoftDeletes was true
+            });
+        }
     }
 };